1_Introduction/react_agent_basic.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv
from langchain.agents import initialize_agent, tool
from langchain_community.tools import TavilySearchResults
import datetime

load_dotenv()

llm = ChatGoogleGenerativeAI(model="gemini-flash-latest")

search_tool = TavilySearchResults(search_depth="basic")

@tool
def get_system_time(format: str = "%Y-%m-%d %h:%M:%S"):
    """
    Return the current date and time in the specified format
    
    :param format: Description
    :type format: str
    """

    current_time = datetime.datetime.now()
    formatted_time = current_time.strftime(format)
    return formatted_time

tools = [search_tool,get_system_time]

agent = initialize_agent(tools=tools, llm=llm, agent="zero-shot-react-description", verbose=True)

agent.invoke("When was the last spacex launch is and how many days ago from today?")

# result = llm.invoke("Give me a twit about todays wether in sri lanka")

# print(result)



